refactor: move checkEmailConfig to server/utils
